alt(chore): Rename from model to mistral
@note

Models are architecture specific and don't always have the same structure which adds a unique level of complexity.

To mitigate this complexity, models should be handled on a case by case basis.

It's probably best to isolate the implementation for each model file format due to the level of specificity and sensitivity.

Each model should follow the specification as a guideline, though differences are expected between parameters, tokenizers, and weights and biases.